Read replica list
